requirements.txt
# FastAPI & Server
fastapi==0.115.13
uvicorn[standard]==0.34.3
python-multipart==0.0.20

# Pydantic
pydantic==2.11.7
pydantic-settings==2.10.1

# Database & Cache  
asyncpg==0.30.0
pgvector>=0.8.0
redis==6.2.0
alembic==1.13.1

# AI & ML (focused on embeddings and search)
openai==1.91.0
numpy==2.0.2
sentence-transformers==4.1.0
tenacity==9.1.2
tiktoken==0.9.0
langchain-text-splitters>=0.3.0

# Search
bm25s==0.2.13
PyStemmer==2.2.0.1
stopwordsiso==0.6.1

# Production
slowapi==0.1.9
circuitbreaker==2.1.3
aiofiles==24.1.0
httpx[http2]==0.28.1
PyJWT[crypto]==2.9.0

# Development
pytest==8.3.2
pytest-asyncio==0.23.8

# Data Import
aiohttp==3.11.11
beautifulsoup4==4.12.3
